requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.29
pydantic==2.7.0
pydantic-settings==2.2.1
alembic==1.13.1
python-multipart==0.0.9
